style: comply to the norm
