+++ include/writer.h
@@ -6,7 +6,7 @@
 /*   By: bdevessi <baptiste@devessier.fr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/16 20:04:19 by bdevessi          #+#    #+#             */
-/*   Updated: 2021/03/17 10:35:11 by bdevessi         ###   ########.fr       */
+/*   Updated: 2021/03/22 13:50:07 by bdevessi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,11 +31,18 @@ typedef struct			s_writer
 }						t_writer;
 
 t_writer				create_writer(int fd, char *filename);
-t_writer				create_writer_breaker(int fd, char *filename, size_t line_break, char *new_line);
+t_writer				create_writer_breaker(int fd
+	, char *filename
+	, size_t line_break
+	, char *new_line);
 t_writer				create_writer_empty();
 
-void					writer_write(t_writer *writer, char *to_write, size_t length);
-void					writer_pad(t_writer *writer, char pad_char, size_t count);
+void					writer_write(t_writer *writer
+	, char *to_write
+	, size_t length);
+void					writer_pad(t_writer *writer
+	, char pad_char
+	, size_t count);
 void					writer_flush(t_writer *writer);
 void					writer_close(t_writer *writer);
