+++ src.mk
@@ -3,8 +3,12 @@ SRCS += srcs/algo_rotate.c
 SRCS += srcs/args.c
 SRCS += srcs/base64.c
 SRCS += srcs/base64_algo.c
+SRCS += srcs/base64_algo_decode.c
+SRCS += srcs/base64_algo_decode_utils.c
+SRCS += srcs/base64_algo_encode.c
 SRCS += srcs/base64_decode.c
 SRCS += srcs/base64_encode.c
+SRCS += srcs/base64_utils.c
 SRCS += srcs/digest.c
 SRCS += srcs/digest_exec.c
 SRCS += srcs/hexa.c
@@ -15,6 +19,7 @@ SRCS += srcs/md5_algo_utils.c
 SRCS += srcs/md5_print.c
 SRCS += srcs/open.c
 SRCS += srcs/reader.c
+SRCS += srcs/reader_read.c
 SRCS += srcs/repl.c
 SRCS += srcs/repl_parser.c
 SRCS += srcs/repl_parser_context.c
@@ -46,3 +51,6 @@ SRCS += srcs/sha512_print.c
 SRCS += srcs/ssl.c
 SRCS += srcs/usage.c
 SRCS += srcs/writer.c
+SRCS += srcs/writer_close.c
+SRCS += srcs/writer_flush.c
+SRCS += srcs/writer_write.c
