+++ srcs/base64.c
@@ -6,15 +6,13 @@
 /*   By: bdevessi <baptiste@devessier.fr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/16 17:17:14 by bdevessi          #+#    #+#             */
-/*   Updated: 2021/03/17 13:01:31 by bdevessi         ###   ########.fr       */
+/*   Updated: 2021/03/22 13:02:17 by bdevessi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 #include "ssl.h"
 #include "base64.h"
-#include "writer.h"
-#include "open.h"
 
 t_arg		g_base64_arguments[] = {
 	{
@@ -71,37 +69,14 @@ void		bind_base64_args(t_context *ctx)
 void		base64_log_command_usage(t_context *ctx)
 {
 	(void)ctx;
-
 	ft_putf_fd(STDERR_FILENO
 		, "usage: ft_ssl base64 [-ed] [-b num] [-i in_file] [-o out_file]\n");
 }
 
-void						base64_cmd(t_context *ctx)
+void		base64_cmd(t_context *ctx)
 {
 	if (ctx->algo_ctx.base64.is_encoding)
 		base64_encode_cmd(ctx);
 	else
 		base64_decode_cmd(ctx);
-}
-
-t_writer					base64_create_writer(t_context *ctx
-	, char *output_file, size_t line_break)
-{
-	int				file_fd;
-	char			*filename;
-
-	if (output_file == NULL)
-	{
-		file_fd = STDOUT_FILENO;
-		filename = "stdout";
-	}
-	else
-	{
-		if ((file_fd = open_write_file(ctx, (char *)output_file)) == -1)
-			return (create_writer_empty());
-		filename = (char *)output_file;
-	}
-	if (line_break == 0)
-		return (create_writer(file_fd, filename));
-	return (create_writer_breaker(file_fd, filename, line_break, "\n"));
 }
